requirements.txt
boto3==1.34.149
channels
channels_redis==4.2.0
celery==5.4.0
django-cors-headers==4.4.0
Django==5.0.7
django-channels==0.7.0 
django_celery_beat==2.6.0
django_celery_results
django-storages==1.14.4
django_redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt ==5.3.1
pillow
psycopg2_binary
psycopg2
python-dotenv==1.0.1
requests
uvicorn==0.30.4
websockets
atproto